+++ db/migrate/20180628024407_change_character_column_name.rb
@@ -1,0 +1,5 @@
+class ChangeCharacterColumnName < ActiveRecord::Migration[5.2]
+  def change
+    rename_column :characters, :name, :title
+  end
+end