+++ db/schema.rb
@@ -10,13 +10,13 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2018_06_27_232746) do
+ActiveRecord::Schema.define(version: 2018_06_28_024407) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
   create_table "characters", force: :cascade do |t|
-    t.string "name"
+    t.string "title"
     t.integer "rank"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false