+++ spec/factories/character.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :character do
+    title { Faker::Ancient.hero }
+    rank  { Faker::Number.between(1, 30) }
+  end
+end