+++ spec/models/character_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper.rb'
+
+describe Character do
+  it "has a valid factory"
+  it "is invalid without a title"
+  it "is invalid without a rank"
+  it "returns a title"
+  it "returns a rank"
+end